+++ hutool-http/src/main/java/cn/hutool/http/webservice/SoapClient.java
@@ -1,0 +1,348 @@
+package cn.hutool.http.webservice;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.nio.charset.Charset;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+
+import javax.xml.XMLConstants;
+import javax.xml.namespace.QName;
+import javax.xml.soap.MessageFactory;
+import javax.xml.soap.Name;
+import javax.xml.soap.SOAPBodyElement;
+import javax.xml.soap.SOAPElement;
+import javax.xml.soap.SOAPException;
+import javax.xml.soap.SOAPHeader;
+import javax.xml.soap.SOAPHeaderElement;
+import javax.xml.soap.SOAPMessage;
+
+import cn.hutool.core.map.MapUtil;
+import cn.hutool.core.util.CharsetUtil;
+import cn.hutool.core.util.StrUtil;
+import cn.hutool.core.util.XmlUtil;
+import cn.hutool.http.HttpRequest;
+
+/**
+ * SOAP客户端
+ * 
+ * @author looly
+ * @since 4.5.4
+ */
+public class SoapClient {
+
+	/** XML消息体的Content-Type */
+	private static final String TEXT_XML_CONTENT_TYPE = "text/xml;charset=";
+
+	/** 请求的URL地址 */
+	private String url;
+	/** 编码 */
+	private Charset charset = CharsetUtil.CHARSET_UTF_8;
+	/** SOAP消息 */
+	private SOAPMessage message;
+	/** 消息方法节点 */
+	private SOAPBodyElement methodEle;
+
+	/**
+	 * 创建SOAP客户端，默认使用soap1.1版本协议
+	 * 
+	 * @param url WS的URL地址
+	 * @return {@link SoapClient}
+	 */
+	public static SoapClient create(String url) {
+		return new SoapClient(url);
+	}
+
+	/**
+	 * 创建SOAP客户端
+	 * 
+	 * @param url WS的URL地址
+	 * @param protocol 协议，见{@link SoapProtocol}
+	 * @return {@link SoapClient}
+	 */
+	public static SoapClient create(String url, SoapProtocol protocol) {
+		return new SoapClient(url, protocol);
+	}
+
+	/**
+	 * 构造，默认使用soap1.1版本协议
+	 * 
+	 * @param url WS的URL地址
+	 */
+	public SoapClient(String url) {
+		this(url, SoapProtocol.SOAP_1_1);
+	}
+
+	/**
+	 * 构造
+	 * 
+	 * @param url WS的URL地址
+	 * @param protocol 协议版本，见{@link SoapProtocol}
+	 */
+	public SoapClient(String url, SoapProtocol protocol) {
+		this.url = url;
+		init(protocol);
+	}
+
+	/**
+	 * 初始化
+	 * 
+	 * @param protocol 协议版本枚举，见{@link SoapProtocol}
+	 * @return this
+	 */
+	public SoapClient init(SoapProtocol protocol) {
+		// 创建消息工厂
+		MessageFactory factory;
+		try {
+			factory = MessageFactory.newInstance(protocol.getValue());
+			// 根据消息工厂创建SoapMessage
+			this.message = factory.createMessage();
+		} catch (SOAPException e) {
+			throw new SoapRuntimeException(e);
+		}
+
+		return this;
+	}
+
+	/**
+	 * 设置编码
+	 * 
+	 * @param charset 编码
+	 * @return this
+	 */
+	public SoapClient setCharset(Charset charset) {
+		this.charset = charset;
+		return this;
+	}
+
+	/**
+	 * 设置Webservice请求地址
+	 * 
+	 * @param url Webservice请求地址
+	 * @return this
+	 */
+	public SoapClient setUrl(String url) {
+		this.url = url;
+		return this;
+	}
+
+	/**
+	 * 设置头信息
+	 * 
+	 * @param name 头信息标签名
+	 * @return this
+	 */
+	public SoapClient setHeader(QName name) {
+		return setHeader(name, null, null, null, null);
+	}
+
+	/**
+	 * 设置头信息
+	 * 
+	 * @param name 头信息标签名
+	 * @param actorURI 中间的消息接收者
+	 * @param roleUri Role的URI
+	 * @param mustUnderstand 标题项对于要对其进行处理的接收者来说是强制的还是可选的
+	 * @param relay relay属性
+	 * @return this
+	 */
+	public SoapClient setHeader(QName name, String actorURI, String roleUri, Boolean mustUnderstand, Boolean relay) {
+		SOAPHeader header;
+		SOAPHeaderElement ele;
+		try {
+			header = this.message.getSOAPHeader();
+			ele = header.addHeaderElement(name);
+			if (StrUtil.isNotBlank(roleUri)) {
+				ele.setRole(roleUri);
+			}
+			if (null != relay) {
+				ele.setRelay(relay);
+			}
+		} catch (SOAPException e) {
+			throw new SoapRuntimeException(e);
+		}
+
+		if (StrUtil.isNotBlank(actorURI)) {
+			ele.setActor(actorURI);
+		}
+		if (null != mustUnderstand) {
+			ele.setMustUnderstand(mustUnderstand);
+		}
+
+		return this;
+	}
+
+	/**
+	 * 设置请求方法
+	 * 
+	 * @param name 方法名及其命名空间
+	 * @param params 参数
+	 * @param useMethodPrefix 是否使用方法的命名空间前缀
+	 * @return this
+	 */
+	public SoapClient setMethod(Name name, Map<String, String> params, boolean useMethodPrefix) {
+		return setMethod(new QName(name.getURI(), name.getLocalName(), name.getPrefix()), params, useMethodPrefix);
+	}
+
+	/**
+	 * 设置请求方法
+	 * 
+	 * @param name 方法名及其命名空间
+	 * @param params 参数
+	 * @param useMethodPrefix 是否使用方法的命名空间前缀
+	 * @return this
+	 */
+	public SoapClient setMethod(QName name, Map<String, String> params, boolean useMethodPrefix) {
+		setMethod(name);
+		final String prefix = name.getPrefix();
+		final SOAPBodyElement methodEle = this.methodEle;
+		for (Entry<String, String> entry : MapUtil.wrap(params)) {
+			setParam(methodEle, entry.getKey(), entry.getValue(), prefix);
+		}
+
+		return this;
+	}
+
+	/**
+	 * 设置请求方法
+	 * 
+	 * @param name 方法名
+	 * @return this
+	 */
+	public SoapClient setMethod(String methodName) {
+		return setMethod(methodName, XMLConstants.NULL_NS_URI);
+	}
+
+	/**
+	 * 设置请求方法<br>
+	 * 方法名自动识别前缀，前缀和方法名使用“:”分隔
+	 * 
+	 * @param name 方法名（可有前缀也可无）
+	 * @return this
+	 */
+	public SoapClient setMethod(String methodName, String namespaceURI) {
+		final List<String> methodNameList = StrUtil.split(methodName, ':');
+		final QName qName;
+		if (2 == methodNameList.size()) {
+			qName = new QName(namespaceURI, methodNameList.get(1), methodNameList.get(0));
+		} else {
+			qName = new QName(namespaceURI, methodName);
+		}
+		return setMethod(qName);
+	}
+
+	/**
+	 * 设置请求方法
+	 * 
+	 * @param name 方法名及其命名空间
+	 * @return this
+	 */
+	public SoapClient setMethod(QName name) {
+		try {
+			this.methodEle = this.message.getSOAPBody().addBodyElement(name);
+		} catch (SOAPException e) {
+			throw new SoapRuntimeException(e);
+		}
+
+		return this;
+	}
+
+	/**
+	 * 设置方法参数，使用方法的前缀
+	 * 
+	 * @param name 参数名
+	 * @param value 参数值
+	 * @return this
+	 */
+	public SoapClient setParam(String name, String value) {
+		return setParam(name, value, true);
+	}
+
+	/**
+	 * 设置方法参数
+	 * 
+	 * @param name 参数名
+	 * @param value 参数值
+	 * @param useMethodPrefix 是否使用方法的命名空间前缀
+	 * @return this
+	 */
+	public SoapClient setParam(String name, String value, boolean useMethodPrefix) {
+		setParam(this.methodEle, name, value, useMethodPrefix ? this.methodEle.getPrefix() : null);
+		return this;
+	}
+
+	/**
+	 * 获取SOAP请求消息
+	 * 
+	 * @param pretty 是否格式化
+	 * @return 消息字符串
+	 */
+	public String getMsgStr(boolean pretty) {
+		final ByteArrayOutputStream out = new ByteArrayOutputStream();
+		try {
+			this.message.writeTo(out);
+		} catch (SOAPException | IOException e) {
+			throw new SoapRuntimeException(e);
+		}
+		return pretty ? XmlUtil.format(out.toString()) : out.toString();
+	}
+
+	/**
+	 * 执行Webservice请求，既发送SOAP内容
+	 * 
+	 * @return 返回结果
+	 */
+	public String execute() {
+		return execute(false);
+	}
+
+	/**
+	 * 执行Webservice请求，既发送SOAP内容
+	 * 
+	 * @param pretty 是否格式化
+	 * @return 返回结果
+	 */
+	public String execute(boolean pretty) {
+		String res = HttpRequest.post(this.url)//
+				.setFollowRedirects(true).contentType(getXmlContentType())//
+				.body(getMsgStr(false))//
+				.execute()//
+				.body();
+
+		return pretty ? XmlUtil.format(res) : res;
+	}
+
+	// -------------------------------------------------------------------------------------------------------- Private method start
+	/**
+	 * 获取请求的Content-Type，附加编码信息
+	 * 
+	 * @return 请求的Content-Type
+	 */
+	private String getXmlContentType() {
+		return TEXT_XML_CONTENT_TYPE.concat(this.charset.toString());
+	}
+
+	/**
+	 * 设置方法参数
+	 * 
+	 * @param methodEle 方法节点
+	 * @param name 参数名
+	 * @param value 参数值
+	 * @param prefix 命名空间前缀
+	 */
+	private static void setParam(SOAPBodyElement methodEle, String name, String value, String prefix) {
+		final SOAPElement childEle;
+		try {
+			if (StrUtil.isNotBlank(prefix)) {
+				childEle = methodEle.addChildElement(name, prefix);
+			} else {
+				childEle = methodEle.addChildElement(name);
+			}
+		} catch (SOAPException e) {
+			throw new SoapRuntimeException(e);
+		}
+		childEle.setValue(value);
+	}
+	// -------------------------------------------------------------------------------------------------------- Private method end
+}